Add transport channel creation property support

+++ components/org.wso2.carbon.sp.jobmanager.core/src/main/java/org/wso2/carbon/sp/jobmanager/core/appcreator/SPSiddhiAppCreator.java
@@ -30,6 +30,7 @@ import org.wso2.carbon.sp.jobmanager.core.topology.PublishingStrategyDataHolder;
 import org.wso2.carbon.sp.jobmanager.core.topology.SiddhiQueryGroup;
 import org.wso2.carbon.sp.jobmanager.core.topology.SubscriptionStrategyDataHolder;
 import org.wso2.carbon.sp.jobmanager.core.util.ResourceManagerConstants;
+import org.wso2.carbon.sp.jobmanager.core.util.SiddhiTopologyCreatorConstants;
 import org.wso2.carbon.sp.jobmanager.core.util.TransportStrategy;
 import org.wso2.siddhi.core.exception.SiddhiAppCreationException;
 
@@ -46,6 +47,7 @@ import java.util.concurrent.TimeUnit;
  */
 public class SPSiddhiAppCreator extends AbstractSiddhiAppCreator {
     private static final Logger log = Logger.getLogger(SPSiddhiAppCreator.class);
+    private static final int TIMEOUT = 120;
 
     @Override
     protected List<SiddhiQuery> createApps(String siddhiAppName, SiddhiQueryGroup queryGroup) {
@@ -111,15 +113,28 @@ public class SPSiddhiAppCreator extends AbstractSiddhiAppCreator {
     }
 
     private void createTopicPartitions(Map<String, Integer> topicParallelismMap) {
-        int timeout = 120;
+        ZkUtils zkUtils;
+        String[] bootstrapServerURLs = null;
+        SafeZkClient safeZkClient;
         String bootstrapServerURL = ServiceDataHolder.getDeploymentConfig().getBootstrapURLs();
-        String[] bootstrapServerURLs = bootstrapServerURL.replaceAll("\\s+", "").split(",");
         String zooKeeperServerURL = ServiceDataHolder.getDeploymentConfig().getZooKeeperURLs();
-        String[] zooKeeperServerURLs = zooKeeperServerURL.replaceAll("\\s+", "").split(",");
 
-        boolean isSecureKafkaCluster = false;
-        SafeZkClient safeZkClient = new SafeZkClient();
-        ZkUtils zkUtils = safeZkClient.createZkClient(zooKeeperServerURLs, isSecureKafkaCluster);
+        if (zooKeeperServerURL == null) {
+            throw new SiddhiAppCreationException("ZooKeeper URLs are not provided " +
+                    "in deployment.yaml under deployment.config. Hence cannot check existence of topics.");
+        } else {
+            String[] zooKeeperServerURLs = zooKeeperServerURL.replaceAll("\\s+", "").split(",");
+            boolean isSecureKafkaCluster = false;
+            safeZkClient = new SafeZkClient();
+            zkUtils = safeZkClient.createZkClient(zooKeeperServerURLs, isSecureKafkaCluster);
+        }
+
+        if (bootstrapServerURL != null) {
+            bootstrapServerURLs = bootstrapServerURL.replaceAll("\\s+", "").split(",");
+        } else if (transportChannelCreationEnabled) {
+            throw new SiddhiAppCreationException("Bootstrap server URLs  are not provided " +
+                    "in deployment.yaml under deployment.config. Hence cannot create required topics.");
+        }
 
         Properties topicConfig = new Properties();
         for (Map.Entry<String, Integer> entry : topicParallelismMap.entrySet()) {
@@ -132,33 +147,50 @@ public class SPSiddhiAppCreator extends AbstractSiddhiAppCreator {
                     (new SafeKafkaInvoker()).addKafkaPartition(zkUtils, topic, partitions);
                     log.info("Added " + partitions + " partitions to topic " + topic);
                 } else if (existingPartitions > partitions) {
-                    log.info("Topic " + topic + " has higher number of partitions than expected partition count. Hence"
-                            + " have to delete the topic and recreate with " + partitions + "partitions.");
-                    AdminUtils.deleteTopic(zkUtils, topic);
-                    long startTime = System.currentTimeMillis();
-                    while (AdminUtils.topicExists(zkUtils, topic)) {
-                        try {
-                            TimeUnit.SECONDS.sleep(1);
-                            if (System.currentTimeMillis() - startTime > timeout * 1000L) {
-                                break;
+                    if (transportChannelCreationEnabled) {
+                        log.info("Topic " + topic + " has higher number of partitions than expected partition count. Hence"
+                                + " have to delete the topic and recreate with " + partitions + "partitions.");
+                        AdminUtils.deleteTopic(zkUtils, topic);
+                        long startTime = System.currentTimeMillis();
+                        while (AdminUtils.topicExists(zkUtils, topic)) {
+                            try {
+                                TimeUnit.SECONDS.sleep(1);
+                                if (System.currentTimeMillis() - startTime > TIMEOUT * 1000L) {
+                                    break;
+                                }
+                            } catch (InterruptedException e) {
+                                Thread.currentThread().interrupt();
                             }
-                        } catch (InterruptedException e) {
-                            Thread.currentThread().interrupt();
                         }
-                    }
-                    if (!AdminUtils.topicExists(zkUtils, topic)) {
-                        (new SafeKafkaInvoker()).createKafkaTopic(bootstrapServerURLs, zkUtils, topicConfig, topic,
-                                partitions);
-                        log.info("Created topic " + topic + "with " + partitions + " partitions.");
+                        if (!AdminUtils.topicExists(zkUtils, topic)) {
+                            (new SafeKafkaInvoker()).createKafkaTopic(bootstrapServerURLs, zkUtils, topicConfig, topic,
+                                    partitions);
+                            log.info("Created topic " + topic + " with " + partitions + " partitions.");
+                        } else {
+                            throw new SiddhiAppCreationException("Topic " + topic + " deletion failed. Hence Could not "
+                                    + "create new topic to facilitate new partitions."
+                            );
+                        }
                     } else {
-                        throw new SiddhiAppCreationException("Topic " + topic + " deletion failed. Hence Could not "
-                                + "create new topic to facilitate new partitions."
-                        );
+                        throw new SiddhiAppCreationException("Number of partitions in the existing topic has higher " +
+                                "number of partitions than the expected count. Hence need to delete and recreate " +
+                                "topic with " + partitions + "partitions.  User has disabled topic creation by " +
+                                "setting"  + SiddhiTopologyCreatorConstants.TRANSPORT_CHANNEL_CREATION_IDENTIFIER +
+                                " property to false. Hence new topics can't be created and Siddhi App deployemnt " +
+                                "will be aborted.");
                     }
                 }
             } else {
-                (new SafeKafkaInvoker()).createKafkaTopic(bootstrapServerURLs, zkUtils, topicConfig, topic, partitions);
-                log.info("Created topic " + topic + "with " + partitions + " partitions.");
+                if (transportChannelCreationEnabled) {
+                    new SafeKafkaInvoker().createKafkaTopic(bootstrapServerURLs, zkUtils, topicConfig, topic,
+                            partitions);
+                    log.info("Created topic " + topic + " with " + partitions + " partitions.");
+                } else {
+                    throw new SiddhiAppCreationException("Topic " + topic + " creation failed. User has disabled " +
+                            "topic creation by setting " +
+                            SiddhiTopologyCreatorConstants.TRANSPORT_CHANNEL_CREATION_IDENTIFIER +
+                            " property to false. Hence Siddhi App deployment will be aborted.");
+                }
             }
         }
         safeZkClient.closeClient();
